requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.4.1
anyio==4.8.0
attrs==24.2.0
av==12.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.9
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.8.0
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.1
pyarrow==17.0.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.22
pydeck==0.9.1
pyee==12.0.0
pyftpdlib==2.0.1
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.38.0
streamlit-autorefresh==1.0.1
streamlit-webrtc==0.47.9
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.1.3
XlsxWriter==3.2.0
